+++ src/components/v2/sections/index.js
@@ -3,3 +3,4 @@ export { default as PageSection } from "./PageSection";
 export { default as SectionDivider } from "./SectionDivider";
 export { default as SectionHeading } from "./SectionHeading";
 export { default as SectionPanel } from "./SectionPanel";
+export { default as SkillsSection } from "./SkillsSection";